Continue moving viewState away from renderer.js
