+++ src/components/viewState/model.js
@@ -1,0 +1,18 @@
+const viewState = {
+  canvasWidth: 800,
+  canvasHeight: 600,
+  dfMapData: false,
+  dragged: false,
+  imageX: 0,
+  imageY: 0,
+  originalImgHeight: 0,
+  originalImgWidth: 0,
+  imgWidth: 0,
+  imgHeight: 0,
+  clickX: 0,
+  clickY: 0,
+  idx: 0,
+  scale: 0
+}
+
+export default viewState
