+++ src/components/viewState/zoom-to.js
@@ -1,0 +1,44 @@
+/**
+   * Zoom to a POI
+   * layer - layer to zoom to
+   * scale - zoom scale (zoom amount)
+   * xTile - xTile to zoom to
+   * yTile - yTile to zoom to
+   */
+export default function (viewState) {
+  return (layer, pscale, xTile, yTile) => {
+    // find the desired layer
+    let found = false
+    let curLayer
+
+    const dfMapData = viewState.dfMapData || {}
+    const { mapData } = dfMapData
+    const { canvasWidth, canvasHeight } = viewState
+
+    if (!mapData) {
+      return
+    }
+
+    for (let i = 0; i < mapData.length; i++) {
+      curLayer = mapData[i]
+      if (curLayer.depth === layer) {
+        found = true
+        break
+      }
+    }
+
+    if (!found) {
+      return
+    }
+
+    viewState.idx = curLayer.index
+    viewState.scale = pscale
+
+    viewState.imgWidth = viewState.originalImgWidth * viewState.scale
+    viewState.imgHeight = viewState.originalImgHeight * viewState.scale
+
+    console.log('[Zoom To]', { windowWidth: canvasWidth, tileWidth: dfMapData.tileWidth, scale: viewState.scale, xTile })
+    viewState.imageX = canvasWidth / 2 - dfMapData.tileWidth * viewState.scale * xTile + dfMapData.tileWidth / 2 * viewState.scale
+    viewState.imageY = canvasHeight / 2 - dfMapData.tileHeight * viewState.scale * yTile + dfMapData.tileHeight / 2 * viewState.scale
+  }
+}
