+++ src/map-viewer.js
@@ -4,23 +4,10 @@ import setupStartingMap from './components/setup-starting-map'
 import rewriteMapLinks from './components/rewrite-map-links'
 import dragAndDrop from './components/drag-and-drop'
 import userInputs from './components/user-inputs'
+import zoomTo from './components/viewState/zoom-to'
+import viewState from './components/viewState/model'
 
 const p5 = require('p5')
-
-const viewState = {
-  dfMapData: false,
-  dragged: false,
-  imageX: 0,
-  imageY: 0,
-  originalImgHeight: 0,
-  originalImgWidth: 0,
-  imgWidth: 0,
-  imgHeight: 0,
-  clickX: 0,
-  clickY: 0,
-  idx: 0,
-  scale: 0
-}
 
 function setup () {
   let mapRenderer
@@ -30,13 +17,13 @@ function setup () {
     window.p5 = p5
     Object.assign(window, parser)
     Object.assign(window, mapRenderer)
-    userInputs(window, mapRenderer)
+    userInputs(window, { viewState })
   }
 
   if (typeof document !== 'undefined') {
-    rewriteMapLinks(document, mapRenderer)
-    setupStartingMap(document, mapRenderer)
-    dragAndDrop(document, window, mapRenderer)
+    rewriteMapLinks(document, { setMapByURL: mapRenderer.setMapByURL })
+    setupStartingMap(document, { setMapByURL: mapRenderer.setMapByURL, zoomTo: zoomTo(viewState) })
+    dragAndDrop(document, window, { viewState })
   }
 }
